rename disconnect component to make it more obvious
