+++ apps/hosted-widget/pages/[hash].tsx
@@ -10,7 +10,7 @@ import Image from "next/image";
 import { useRouter } from "next/router";
 import { useMemo } from "react";
 
-import StartupDisconnect from "../src/components/startup-disconnect/StartupDisconnect";
+import DemoWalletDisconnect from "../src/components/startup-disconnect/DemoWalletDisconnect";
 import useAnalyticsBrowser from "../src/hooks/useAnalyticsBrowser";
 import useFontLoader from "../src/hooks/useFontLoader";
 import useLoadFromIPFS from "../src/hooks/useLoadFromIPFS";
@@ -47,7 +47,7 @@ const IPFSWidgetPage: NextPage = () => {
 
   return (
     <WagmiProviders>
-      <StartupDisconnect />
+      <DemoWalletDisconnect />
       {showLoader ? (
         <Box
           sx={{
